Fixup, test steady batched lookup type handling

+++ downstream/_auxlib/_bit_floor32.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 from ._bitlen32 import bitlen32
 
 
@@ -6,5 +8,5 @@ def bit_floor32(n: int) -> int:
 
     If zero, returns zero.
     """
-    mask = 1 << bitlen32(n >> 1)
+    mask = np.uint64(1) << bitlen32(n >> 1)
     return n & mask
